third_party/blink/web_tests/external/wpt/css/css-shapes/shape-functions/shape-function-computed.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Property clip-path value 'shape(from center, curve by 20px 20px with 10px 30px from end / 12px 32px from start)'
  assert_equals: expected "shape(from 50% 50%, curve by 20px 20px with 10px 30px from end / 12px 32px from start)" but got "shape(from 50% 50%, curve by 20px 20px with 10px 30px from end / 12px 32px)"
[FAIL] Property clip-path value 'shape(from 20px 40px, curve to top right with 10px 30px from origin / 12px 32px from origin)'
  assert_equals: expected "shape(from 20px 40px, curve to 100% 0% with 10px 30px from origin / 12px 32px from origin)" but got "shape(from 20px 40px, curve to 100% 0% with 10px 30px / 12px 32px)"
Harness: the test ran to completion.
